refactor!: change markup to djot
Fresh new start again! This commit scraps the Pandoc and Pandoc Markdown
dependency in favour of using Djot. Djot has (almost) all the features
we need for fully fledged academic writing, and I'm co-maintaining the
Rust parser for it. It is therefore a perfect fit for this project and
something that should help distinguish it from other similar
aspirations.

This commit is huge, and for good reason. A lot of dependencies have
been replaced or removed, and the project structure has been drastically
simplified. I'm going for more of a KISS approach, to tackle the
inevitable complexity that no doubt will appear with added features. A
breath of fresh air, no doubt.